Include all valid certificates in config tool
In the previous version, the GetSSLCertificatesAndNames method uses the FindByKeyUsage FindType to locate valid certs for SSL trust. This does not adequately allow all possible certificates which a system might use for SSL to be added to the configuration file. By changing to "FindByTimeValid," we instead list all currently-valid certificates on the system. This does put additional onus on the administrator to know which certificate is the appropriate certificate - but should reduce head-scratching when their certificate isn't listed.

See SourceForge support request https://sourceforge.net/p/seb/support-requests/55/ for background on the issue in my environment.