Re-applied "Merge pull request #61 from t00/feature/dowload_fix"
Preventing the cancelling of the download dialog when SEB configuration is in progress (password dialog is shown or reconfiguration error is still displayed)
Downloading seb:// or sebs:// settings in the SEB opens additional dialogs - added an option to disable the download window enabling background download of the seb file (no UI option to configure it).
See https://github.com/SafeExamBrowser/seb-win/issues/58
and https://github.com/SafeExamBrowser/seb-win/pull/61

which was lost because it wasn't pushed into the seb2 browser app repository.